More numbers
Add more Korean numbers and add 1-10 in Sino-Korean numbers.

I'm not sure about mixing the two, it might be better to only test
one number system at a time.

Make service worker update the cache for the next page load when fetching pages.

This shouldn't really be needed because all the pages could be fetched when a
new version of the service worker is loaded. But I'm not sure the service
worker is updating properly at the moment - I don't change the URL of the
service worker script, which might get in the way of this?